Add files via upload
Hello, here is my completion of the Front-end Challenge! I'm submitting a pull request with the files, but I've also deployed it on GitPages in my own repository at: https://stevenrichardshirley.github.io/headstorm-challenge/

Any feedback would be great! I tried to draw up a good UI, and the form works to print out contact information in the console.